Fix build

+++ source/MRViewer/MRRenderMeshObject.cpp
@@ -727,7 +727,7 @@ RenderBufferRef<unsigned> RenderMeshObject::loadFaceSelectionTextureBuffer_()
 
     const auto& selection = objMesh_->getSelectedFaces().m_bits;
     const unsigned* selectionData = ( unsigned* )selection.data();
-    tbb::parallel_for( tbb::blocked_range<int>( 0, buffer.size() ), [&] ( const tbb::blocked_range<int>& range )
+    tbb::parallel_for( tbb::blocked_range<int>( 0, (int)buffer.size() ), [&] ( const tbb::blocked_range<int>& range )
     {
         for ( int r = range.begin(); r < range.end(); ++r )
         {
